rename Tasks::Builder.ship and move Ufo::Ecs::Service into file
