+++ lib/ufo/ecs.rb
@@ -1,0 +1,5 @@
+module Ufo
+  module Ecs
+    autoload :Service, 'ufo/ecs/service'
+  end
+end
